requirements.txt
pvporcupine
pvrecorder
google-cloud-speech
edge-tts
pygame
pynput
pyyaml
httpx
websockets
python-dotenv
pystray
Pillow
